Update filtering functionality, tag filter state
Barring some issues with visuals and rerendering, filtering of events is
functional and works as intended. Events are ordered by closeness to
filters. The more filters that match the user-chosen selection, the
higher rank the event will have.

Futher work needs to be done on managing state; the aforementioned
visual bugs arise because we rerender the filter bar on every filter
change. This is because we're storing the whole tag filter state, as
well as the expand state for the filter groups, locally in
MainDashboard. This was done to ensure that the filter bar and eventlist
both had access to these parts, and wouldn't reset on filter changes.