Upadted ReadMe.md
Organized code to look a bit cleaner when viewing code on a text editor or editing on github raw format, changed names for students. Need to get the images and url's linked up.

There was a mixup where I added these changes to the staging, reverted the changes and made them official on the Master branch